Restructured components folder
Fixed pathes in main.scss
