+++ components/layout/Box/Box.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { cn } from '../../lib/cn';
+import { cn } from '../../../lib/cn';
 
-import { GenericProps } from '../../types/GenericProps';
+import { GenericProps } from '../../../types/GenericProps';
 
 interface Props extends GenericProps {
   children: React.ReactNode;
