+++ components/profile/Experience/Experience.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import { GenericProps } from '../../types/GenericProps';
-import { Experience as ExperienceType } from '../../types/Experience';
+import { GenericProps } from '../../../types/GenericProps';
+import { Experience as ExperienceType } from '../../../types/Experience';
 
-import { cn } from '../../lib/cn';
-import Headline from '../Headline/Headline';
+import { cn } from '../../../lib/cn';
+import Headline from '../../ui/Headline/Headline';
 
 interface ExperienceProps extends ExperienceType, GenericProps {
   current: boolean;
