+++ components/profile/Skill/Skill.scss
