+++ components/ui/Link/Link.tsx
@@ -1,9 +1,9 @@
 import React, { ReactNode } from 'react';
 import { FiLink } from 'react-icons/fi';
 
-import { GenericProps } from '../../types/GenericProps';
+import { GenericProps } from '../../../types/GenericProps';
 
-import { cn } from '../../lib/cn';
+import { cn } from '../../../lib/cn';
 
 interface Props extends GenericProps {
   children: ReactNode;
