+++ components/ui/Text/Text.tsx
@@ -1,8 +1,8 @@
 import React, { CSSProperties, ReactNode } from 'react';
 
-import { GenericProps } from '../../types/GenericProps';
+import { GenericProps } from '../../../types/GenericProps';
 
-import { cn } from '../../lib/cn';
+import { cn } from '../../../lib/cn';
 
 interface Props extends GenericProps {
   children: ReactNode;
