+++ pages/components/index.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 
-import Grid from '../../components/Grid/Grid';
-import Section from '../../components/Section/Section';
+import Grid from '../../components/layout/Grid/Grid';
+import Section from '../../components/layout/Section/Section';
 import Row from '../../partials/Row/Row';
-import Button from '../../components/Button/Button';
-import Skill from '../../components/Skill/Skill';
-import Experience from '../../components/Experience/Experience';
-import Headline from '../../components/Headline/Headline';
-import Link from '../../components/Link/Link';
+import Button from '../../components/ui/Button/Button';
+import Skill from '../../components/profile/Skill/Skill';
+import Experience from '../../components/profile/Experience/Experience';
+import Headline from '../../components/ui/Headline/Headline';
+import Link from '../../components/ui/Link/Link';
 import Page from '../../partials/Page/Page';
-import Text from '../../components/Text/Text';
+import Text from '../../components/ui/Text/Text';
 
 import { Experience as ExperienceType } from '../../types/Experience';
 import { Skill as SkillType } from '../../types/Skill';
